detect if file exists

+++ install.sh
@@ -18,6 +18,5 @@ echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks
 for file in $files; do
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/.$file $HOME/.$file
+  [[ ! -e $HOME/.$file ]] && ln -s $dir/.$file $HOME/.$file || echo ".$file already exists..."
 done
